Lots of changes
Made it to where the app would no longer log me in based on a log_status i had in app memory. Instead created a session that tracks the log status of a user and shows either a login screen or mainscreen based on that status. Also switched my error checking in the login and signup functions to the newLogin view. This changed created new bugs, i created a new loading indicator and fixed the issue i was having with it. On next commit will fix other issues